Fix admin-frontend test pollution and isolated suite mocks
Remove in-process AdminScriptRunModal tests that failed when
AdminScriptList.test mocks the modal globally. Move coverage to the
isolated suite with confirm-first flow assertions.

Update AdminNestedArrayField isolated tests to mock
AdminFieldRendererCore (actual import) and admin-field-* testIDs.
