+++ admin-frontend/src/isolated/AdminScriptRunModal.isolated.tsx
@@ -4,7 +4,10 @@ import {beforeEach, describe, expect, it, mock} from "bun:test";
 import {renderWithTheme} from "@terreno/ui/src/test-utils";
 import React from "react";
 import type {ReactTestInstance} from "react-test-renderer";
-import {act} from "../../../ui/node_modules/@testing-library/react-native";
+import {
+  act,
+  fireEvent,
+} from "../../../ui/node_modules/@testing-library/react-native";
 import type {AdminApi} from "../types";
 
 // Mock @terreno/ui so the Modal renders its children inline. The real Modal
@@ -84,6 +87,15 @@ const waitTicks = async (ms = 40): Promise<void> => {
   });
 };
 
+const pressDryRun = async (
+  getByTestId: (id: string) => ReactTestInstance
+): Promise<void> => {
+  await act(async () => {
+    fireEvent.press(getByTestId("admin-script-dry-run-button"));
+  });
+  await waitTicks();
+};
+
 describe("AdminScriptRunModal", () => {
   beforeEach(() => {
     runCalls.length = 0;
@@ -107,19 +119,95 @@ describe("AdminScriptRunModal", () => {
     expect(runCalls.length).toBe(0);
   });
 
-  it("auto-starts script when modal opens", async () => {
-    renderWithTheme(
+  it("shows confirm-phase Dry Run, Run, and Cancel buttons when visible", () => {
+    const {getByTestId} = renderWithTheme(
       <AdminScriptRunModal
         api={mockApi}
         baseUrl="/admin"
-        onDismiss={() => {}}
-        scriptName="test-script"
-        visible={true}
+        onDismiss={() => undefined}
+        scriptName="my-script"
+        visible
       />
     );
+
+    expect(getByTestId("admin-script-dry-run-button")).toBeTruthy();
+    expect(getByTestId("admin-script-wet-run-button")).toBeTruthy();
+    expect(getByTestId("admin-script-confirm-cancel-button")).toBeTruthy();
+    expect(runCalls.length).toBe(0);
+  });
+
+  it("calls onDismiss from confirm Cancel without starting a run", async () => {
+    const onDismiss = mock(() => undefined);
+
+    const {getByTestId} = renderWithTheme(
+      <AdminScriptRunModal
+        api={mockApi}
+        baseUrl="/admin"
+        onDismiss={onDismiss}
+        scriptName="my-script"
+        visible
+      />
+    );
+
+    await act(async () => {
+      fireEvent.press(getByTestId("admin-script-confirm-cancel-button"));
+    });
+
+    expect(onDismiss).toHaveBeenCalled();
+    expect(runCalls.length).toBe(0);
+  });
+
+  it("starts a dry run when Dry Run is pressed", async () => {
+    const {getByTestId} = renderWithTheme(
+      <AdminScriptRunModal
+        api={mockApi}
+        baseUrl="/admin"
+        onDismiss={() => undefined}
+        scriptName="my-script"
+        visible
+      />
+    );
+
+    await act(async () => {
+      fireEvent.press(getByTestId("admin-script-dry-run-button"));
+    });
     await waitTicks();
+
     expect(runCalls.length).toBe(1);
-    expect(runCalls[0]).toEqual({name: "test-script", wetRun: true});
+    expect(runCalls[0]).toEqual({name: "my-script", wetRun: false});
+  });
+
+  it("hides the wet-run button when dryRunOnly is true", () => {
+    const {getByTestId, queryByTestId} = renderWithTheme(
+      <AdminScriptRunModal
+        api={mockApi}
+        baseUrl="/admin"
+        dryRunOnly
+        onDismiss={() => undefined}
+        scriptName="my-script"
+        visible
+      />
+    );
+
+    expect(getByTestId("admin-script-dry-run-button")).toBeTruthy();
+    expect(getByTestId("admin-script-confirm-cancel-button")).toBeTruthy();
+    expect(queryByTestId("admin-script-wet-run-button")).toBeNull();
+  });
+
+  it("renders the script description in the confirm phase", () => {
+    const {getByText} = renderWithTheme(
+      <AdminScriptRunModal
+        api={mockApi}
+        baseUrl="/admin"
+        onDismiss={() => undefined}
+        scriptDescription="Migrates legacy records"
+        scriptName="migrate-data"
+        visible
+      />
+    );
+
+    expect(getByText("Migrates legacy records")).toBeTruthy();
+    expect(getByText("migrate-data")).toBeTruthy();
   });
 
   it("renders running phase with progress details", async () => {
@@ -301,7 +389,7 @@ describe("AdminScriptRunModal", () => {
     state.runImpl = () => ({
       unwrap: () => Promise.reject({data: {detail: "Detailed failure"}}),
     });
-    renderWithTheme(
+    const {getByTestId} = renderWithTheme(
       <AdminScriptRunModal
         api={mockApi}
         baseUrl="/admin"
@@ -310,7 +398,7 @@ describe("AdminScriptRunModal", () => {
         visible={true}
       />
     );
-    await waitTicks(200);
+    await pressDryRun(getByTestId);
     expect(runCalls.length).toBe(1);
   });
 
@@ -318,7 +406,7 @@ describe("AdminScriptRunModal", () => {
     state.runImpl = () => ({
       unwrap: () => Promise.reject({data: {title: "Title failure"}}),
     });
-    renderWithTheme(
+    const {getByTestId} = renderWithTheme(
       <AdminScriptRunModal
         api={mockApi}
         baseUrl="/admin"
@@ -327,7 +415,7 @@ describe("AdminScriptRunModal", () => {
         visible={true}
       />
     );
-    await waitTicks(200);
+    await pressDryRun(getByTestId);
     expect(runCalls.length).toBe(1);
   });
 
@@ -335,7 +423,7 @@ describe("AdminScriptRunModal", () => {
     state.runImpl = () => ({
       unwrap: () => Promise.reject(new Error("network boom")),
     });
-    renderWithTheme(
+    const {getByTestId} = renderWithTheme(
       <AdminScriptRunModal
         api={mockApi}
         baseUrl="/admin"
@@ -344,7 +432,7 @@ describe("AdminScriptRunModal", () => {
         visible={true}
       />
     );
-    await waitTicks(200);
+    await pressDryRun(getByTestId);
     expect(runCalls.length).toBe(1);
   });
